latest build

+++ include/Utils.h
@@ -106,7 +106,7 @@ namespace
 	typedef std::size_t UInt;
 	typedef unsigned char Byte;
 	
-	static const auto MAX_THREADS = omp_get_max_threads();
+	static const auto MAX_THREADS = static_cast<UInt>(omp_get_max_threads());
 	auto GetThreads(const UInt elements, const Float weight = Float(1)) NOEXCEPT
 	{
 		const auto load = static_cast<UInt>(Float(elements) * weight);
